+++ CHANGELOG.md
@@ -67,3 +67,4 @@
 - Bugfix, when on first load parcel is not selected.
 - Bugfix, when carrier cannot be changed in back office order view and label is not printed
 - Bugfix, when order is using voucher, recalculate parcel total to match order total(shipping discount excluded)
+- Feature change, remove cache from symfony container compilation(fixes problems for clients which does not have permissions for folder write inside module)